requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.57.0
sentence-transformers==3.3.1
requests==2.32.3
beautifulsoup4==4.12.3
pydantic==2.10.3
numpy==1.26.4
scikit-learn==1.5.2
python-dotenv==1.0.1
httpx==0.28.1
lxml==5.3.0
